+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- store HTTP response metadata in ~/.cache
+
 ## [0.32.0] - 2019-05-28
 
 ### Added